+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to the ld-find-code-refs program will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org).
 
+## [2.0.0] - prerelease
+
+### Changed
+
+- The underlying command line package has been changed for improved ergonomics when configuring options using the command line. Some existing configurations may be invalid, see `ld-find-code-refs --help` for details.
+- Delimiters will no longer provided by default. To reduce the number of "false positive" code references, specify the `--delimiters` option.
+
+### Removed
+
+- The `exclude` command-line option has been removed. Use a `.ldignore` file instead.
+
 ## [1.5.0] - prerelease
 
 ### Added